refactor release steps
